+++ pages/history/comparisons.tsx
@@ -10,8 +10,8 @@ import { SeasonTypeRadios } from 'components/SeasonTypeRadios';
 import { useComparisonFilters } from 'hooks/useComparisonFilters';
 import { usePlayerComparison } from 'hooks/usePlayerComparison';
 import { NextSeo } from 'next-seo';
-import { PlayerNames } from 'pages/api/v2/player/playerSearch';
 import { useCallback, useEffect, useState } from 'react';
+import { PlayerNames } from 'typings/api';
 import { leagueNameToId } from 'utils/leagueHelpers';
 import { query } from 'utils/query';
 import { seasonTypeToTextFriendly } from 'utils/seasonTypeHelpers';
